backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.1.2+cpu --index-url https://download.pytorch.org/whl/cpu
torchvision==0.16.2+cpu --index-url https://download.pytorch.org/whl/cpu
pillow==10.2.0
numpy==1.26.3
pinecone-client==3.0.0
python-dotenv==1.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
requests==2.31.0